Avoid compile problem with std::pow by static_cast of the exponent to int.
git-svn-id: https://svn.dealii.org/trunk@20132 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/codim_one/integrate_log.cc
@@ -156,11 +156,11 @@ int main()
     for (unsigned int k=0; k<=j; k++)
 	exact_integral +=
 	    newton_binomial(j,k)
-	    *pow(a,j-k)
+	    *pow(a,static_cast<int>(j-k))
 	    *(
-		pow(b-a,k+1) / (k+1) * log(b-a)
+	      pow(b-a,static_cast<int>(k)+1) / (k+1) * log(b-a)
 		-
-		pow(b-a,k+1) / pow(k+1, 2)
+	      pow(b-a,static_cast<int>(k)+1) / pow(k+1, 2)
 	     ); 
     deallog << "f(x) = x^" << j << std::endl;
     for (unsigned int i=1; i<13; ++i)
